+++ CHANGELOG.md
@@ -2,16 +2,31 @@
 
 ## [Table of Contents]
 - [Unreleased](#unreleased)
+- [1.0.1](#101---2021-05-12)
 - [1.0.0](#100---2021-05-01)
 
 ## [Unreleased][]
 ### Added
+### Changed
+### Removed
+
+## [1.0.1] - 2021-05-12
+### Added
 - Portable: Additional checks for portable installations.
 - Additional check for biography enabled property to avoid conflicts when manually editing properties.
+- Files: helpers\world_map_menu.js and helpers\world_map_helpers.js.
+- 3 key modifiers for playlist creation supported.
+
 ### Changed
 - Bug fix: refresh points after showing the tooltip on one and changing the selected tracks.
 - Map framework: updated with last changes.
+- Menu: moved all code to helpers\world_map_menu.js.
+- Menu: added entries to edit tags related to playlist creation when using key modifiers + left click.
+- Helpers: moved all code to helpers\world_map_helpers.js.
+- Breaks properties from previous versions!
+
 ### Removed
+
 
 ## [1.0.0] - 2021-05-01
 ### Added
@@ -21,5 +36,6 @@
 
 ### Removed
 
-[Unreleased]: https://github.com/regorxxx/World-Map-SMP/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/regorxxx/World-Map-SMP/compare/v1.0.1...HEAD
+[1.0.1]: https://github.com/regorxxx/World-Map-SMP/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/regorxxx/World-Map-SMP/compare/8df8206...v1.0.0